pyproject.toml
[tool.poetry]
name = "my-first-django"
version = "0.1.0"
description = ""
authors = ["Yebin Kim <rladpqls07@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "my_first_django"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.2"
black = "^22.10.0"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
